requirements.txt
df==0.0.1
Flask==3.0.3
jproperties==2.1.1
librosa==0.10.1
numpy==1.26.4
openai_whisper==20231117
pandas==2.2.2
psutil==5.9.8
pyAudioAnalysis==0.3.14
pydub==0.25.1
Resemblyzer==0.1.4
scikit_learn==1.4.1.post1
simple_diarizer==0.0.13
soundfile==0.12.1
